Master

+++ ITIExaminationSyustem/Controllers/UserController.cs
@@ -1,6 +1,7 @@
 ﻿using ITIExaminationSyustem.Interfaces;
 using ITIExaminationSyustem.Models;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.EntityFrameworkCore;
 using System.Data;
 
 namespace ITIExaminationSyustem.Controllers
@@ -23,13 +24,28 @@ namespace ITIExaminationSyustem.Controllers
         public IActionResult Index()
         {
             List<User> users = _userRepo.GetAll();
+            _userRepo.GetNonAssignedUsers();
             return View(users);
         }
 
         public IActionResult Details(int? id)
         {
-            User fetchedUser = _userRepo.GetById(id.Value);
-            return View(fetchedUser);
+            if (id == null)
+            {
+                return BadRequest();
+            }
+            else
+            {
+                User fetchedUser = _userRepo.GetById(id.Value);
+                if (fetchedUser == null)
+                {
+                    return NotFound();
+                }
+                else
+                {
+                    return View(fetchedUser);
+                }
+            }
         }
 
         public IActionResult Create()
@@ -38,11 +54,12 @@ namespace ITIExaminationSyustem.Controllers
         }
 
         [HttpPost]
-        public IActionResult Create(User user) //image is missing
+        async public Task<IActionResult> Create(User user, IFormFile image)
         {
             if(ModelState.IsValid)
             {
                 _userRepo.Add(user);
+                await _userRepo.AddImage(user, image);
                 return RedirectToAction("Index");
             }
             else
@@ -53,17 +70,32 @@ namespace ITIExaminationSyustem.Controllers
 
         public IActionResult Edit(int? id)
         {
-            User userToEdit = _userRepo.GetById(id.Value);
-            return View(userToEdit);
+            if(id == null)
+            {
+                return BadRequest();
+            }
+            else
+            {
+                User userToEdit = _userRepo.GetById(id.Value);
+                if(userToEdit == null) 
+                { 
+                    return NotFound();
+                }
+                else
+                {
+                    return View(userToEdit);
+                }
+            }
         }
 
         [HttpPost]
-        public IActionResult Edit(User user, int id) //image is missing
+        async public Task<IActionResult> Edit(User user, int id, IFormFile image)
         {
             user.User_Id = id;
             if (ModelState.IsValid)
             {
                 _userRepo.Update(user);
+                await _userRepo.AddImage(user, image);
                 return RedirectToAction("Index");
             }
             else
@@ -72,21 +104,34 @@ namespace ITIExaminationSyustem.Controllers
             }
         }
 
-        public IActionResult Delete(int id) //Completed
+        public IActionResult Delete(int? id)
         {
-            User user = _userRepo.GetById(id);
+            if(id == null)
+            {
+                return BadRequest();
+            }
+            else
+            {
+                User user = _userRepo.GetById(id.Value);
+                if (user == null)
+                {
+                    return NotFound();
+                }
+                else
+                {
+                    if (user.Navigation_Instructor != null)
+                        _instructorRepo.Delete(user.Navigation_Instructor.Instructor_Id);
 
-            if (user.Navigation_Instructor != null)
-                _instructorRepo.Delete(user.Navigation_Instructor.Instructor_Id);
+                    else if (user.Navigation_Student != null)
+                        _studentRepo.Delete(user.Navigation_Student.Student_Id);
 
-            else if (user.Navigation_Student != null)
-                _studentRepo.Delete(user.Navigation_Student.Student_Id);
+                    else if (user.Navigation_Admin != null)
+                        _adminRepo.Delete(user.Navigation_Admin.Admin_Id);
 
-            else if (user.Navigation_Admin != null)
-                _adminRepo.Delete(user.Navigation_Admin.Admin_Id);
-
-            _userRepo.Delete(id);
-            return RedirectToAction("Index");
+                    _userRepo.Delete(id.Value);
+                    return RedirectToAction("Index");
+                }
+            }
         }
     }
 }
